add: migrate

+++ src/repositories/sequelize-mysql/RoomAddChannelRepositoryImpl.ts
@@ -26,9 +26,9 @@ class RoomAddChannelRepositoryImpl
 	@AutoIncrement
 	@Column(DataType.INTEGER)
 	declare id: number;
-	@Column(DataType.BIGINT)
+	@Column(DataType.INTEGER)
 	declare communityId: number;
-	@Column(DataType.BIGINT)
+	@Column(DataType.INTEGER)
 	declare channelId: number;
 
 	async create(data: RoomAddChannelDto): Promise<boolean> {
